ChatListener: Remove and move logic to ClientListener
Taken from versions/1.16.2 branch

This was a small class, and there was really no point in the separation from ClientListener
