+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/libs/ChatLib.kt
@@ -1,7 +1,7 @@
 package com.chattriggers.ctjs.minecraft.libs
 
 import com.chattriggers.ctjs.minecraft.libs.renderer.Renderer
-import com.chattriggers.ctjs.minecraft.listeners.ChatListener
+import com.chattriggers.ctjs.minecraft.listeners.ClientListener
 import com.chattriggers.ctjs.minecraft.objects.message.Message
 import com.chattriggers.ctjs.minecraft.objects.message.TextComponent
 import com.chattriggers.ctjs.minecraft.wrappers.Client
@@ -384,7 +384,7 @@ object ChatLib {
      */
     @JvmStatic
     fun getChatLines(): List<String> {
-        val hist = ChatListener.chatHistory.toMutableList()
+        val hist = ClientListener.chatHistory.toMutableList()
         hist.reverse()
         return hist
     }
