+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/listeners/ClientListener.kt
@@ -1,11 +1,15 @@
 package com.chattriggers.ctjs.minecraft.listeners
 
+import com.chattriggers.ctjs.minecraft.libs.ChatLib
+import com.chattriggers.ctjs.minecraft.libs.EventLib
 import com.chattriggers.ctjs.minecraft.wrappers.Scoreboard
 import com.chattriggers.ctjs.minecraft.wrappers.World
 import com.chattriggers.ctjs.minecraft.wrappers.objects.block.BlockFace
 import com.chattriggers.ctjs.minecraft.wrappers.objects.entity.PlayerMP
 import com.chattriggers.ctjs.minecraft.wrappers.objects.inventory.Item
+import com.chattriggers.ctjs.printToConsole
 import com.chattriggers.ctjs.triggers.TriggerType
+import com.chattriggers.ctjs.utils.Config
 import com.chattriggers.ctjs.utils.kotlin.MCBlockPos
 import io.netty.channel.ChannelDuplexHandler
 import io.netty.channel.ChannelHandlerContext
@@ -14,10 +18,7 @@ import net.minecraft.client.renderer.GlStateManager
 import net.minecraft.entity.player.EntityPlayerMP
 import net.minecraft.network.Packet
 import net.minecraft.util.EnumFacing
-import net.minecraftforge.client.event.DrawBlockHighlightEvent
-import net.minecraftforge.client.event.GuiOpenEvent
-import net.minecraftforge.client.event.GuiScreenEvent
-import net.minecraftforge.client.event.RenderGameOverlayEvent
+import net.minecraftforge.client.event.*
 import net.minecraftforge.event.entity.item.ItemTossEvent
 import net.minecraftforge.event.entity.player.EntityItemPickupEvent
 import net.minecraftforge.event.entity.player.PlayerInteractEvent
@@ -29,9 +30,38 @@ import org.mozilla.javascript.Context
 
 object ClientListener {
     private var ticksPassed: Int = 0
+    val chatHistory = mutableListOf<String>()
+    val actionBarHistory = mutableListOf<String>()
 
     init {
         ticksPassed = 0
+    }
+
+    @SubscribeEvent
+    fun onReceiveChat(event: ClientChatReceivedEvent) {
+        when (EventLib.getType(event)) {
+            in 0..1 -> {
+                // save to chatHistory
+                chatHistory += ChatLib.getChatMessage(event, true)
+                if (chatHistory.size > 1000) chatHistory.removeAt(0)
+
+                // normal Chat Message
+                TriggerType.Chat.triggerAll(ChatLib.getChatMessage(event, false), event)
+
+                // print to console
+                if (Config.printChatToConsole) {
+                    "[CHAT] ${ChatLib.replaceFormatting(ChatLib.getChatMessage(event, true))}".printToConsole()
+                }
+            }
+            2 -> {
+                // save to actionbar history
+                actionBarHistory += ChatLib.getChatMessage(event, true)
+                if (actionBarHistory.size > 1000) actionBarHistory.removeAt(0)
+
+                // action bar
+                TriggerType.ActionBar.triggerAll(ChatLib.getChatMessage(event, false), event)
+            }
+        }
     }
 
     @SubscribeEvent
